pyproject.toml
[tool.poetry]
name = "fact-check"
version = "0.1.0"
description = "Antisemitism fact-checking assistant with OCR + RAG pipeline"
authors = ["Duy Huynh <duy@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
python-multipart = "^0.0.9"
pydantic = "^2.7.0"
pydantic-settings = "^2.1.0"
sqlmodel = "^0.0.22"
psycopg = { extras = ["binary"], version = "^3.2.0" }
httpx = "^0.27.0"
llama-index = "^0.11.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
openai = "^1.40.0"
numpy = "^1.26.0"
scikit-learn = "^1.5.0"
pillow = "^10.3.0"
pytesseract = "^0.3.10"
pdfplumber = "^0.11.0"
rapidocr-onnxruntime = "^1.3.24"
tenacity = "^8.3.0"
python-docx = "^1.1.0"
arq = "^0.25.0"
spacy = "^3.7.0"
google-generativeai = "^0.8.0"
sentence-transformers = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
mypy = "^1.10.0"
ruff = "^0.6.0"
black = "^24.8.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

